Add mnemonics and button icons, add version field to Constants
