+++ src/main/java/tamagotchi/gui/BufferedImageLoader.java
@@ -6,15 +6,15 @@ import java.io.IOException;
 
 public class BufferedImageLoader {
 
-  private BufferedImage image;
+    private BufferedImage image;
 
-  public BufferedImage loadImage(String path) {
-    try {
-      image = ImageIO.read(getClass().getClassLoader().getResource(path));
-      //            image = ImageIO.read(new File(path));
-    } catch (IOException e) {
-      e.printStackTrace();
+    public BufferedImage loadImage(String path) {
+        try {
+            image = ImageIO.read(getClass().getClassLoader().getResource(path));
+            //            image = ImageIO.read(new File(path));
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        return image;
     }
-    return image;
-  }
 }
